+++ src/Sensors/I2C/Loom_MPU6050.h
@@ -45,7 +45,7 @@ public:
 	/// \param[in]	calibrate				Bool | <true> | {true, false} | Whether or not to calibrate at start
 	Loom_MPU6050(
 			byte			i2c_address		= 0x69,
-			int				mux_port		= -1,
+			uint8_t			mux_port		= -1,
 			const char*		module_name		= "MPU6050",
 			bool			calibrate		= true
 		);
